@@ -8,22 +8,24 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
+        "php": "^8.0",
+        "barryvdh/laravel-ide-helper": "^2.10",
         "fideloper/proxy": "^4.2",
         "fruitcake/laravel-cors": "^1.0",
         "google/apiclient": "^2.0",
         "guzzlehttp/guzzle": "^6.3",
         "jeroennoten/laravel-adminlte": "^3.3",
-        "laravel/framework": "^7.0",
+        "laravel/framework": "^8.0",
         "laravel/tinker": "^2.0",
         "rachidlaasri/laravel-installer": "^4.1",
         "seriyyy95/wpconnector-library": "dev-master",
-        "smi2/phpclickhouse": "^1.3"
+        "smi2/phpclickhouse": "^1.3",
+        "spatie/laravel-settings": "^2.1"
     },
     "require-dev": {
         "facade/ignition": "^2.0",
-        "fzaninotto/faker": "^1.9.1",
-        "laravel/ui": "^2.0",
+        "fzaninotto/faker": "*",
+        "laravel/ui": "*",
         "mockery/mockery": "^1.3.1",
         "nunomaduro/collision": "^4.1",
         "phpunit/phpunit": "^8.5"
