@@ -2,19 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+use App\Repositories\GscSitesRepository;
+use App\Repositories\GscTaskRepository;
 use Illuminate\Support\Facades\Auth;
-use Session;
-use Storage;
-use App\OptionsManager;
-use App\GoogleDataLoader;
-use App\GoogleNeedConfigFileException;
-use App\GoogleAPI;
 use App\ClickHousePositions;
-use App\GoogleGscSite;
-use App\GscTask;
-use App\Jobs\GscLoadJob;
-use App\LoadLogger;
+use App\Models\GoogleGscSite;
 
 class GscAccountsController extends Controller
 {
@@ -24,154 +16,33 @@ class GscAccountsController extends Controller
         $this->middleware('demo');
     }
 
-    public function index()
+    public function index(GscSitesRepository $repo)
     {
-        $user = Auth::user();
-        $optionsManager = new OptionsManager();
-        $optionsManager->setUser($user->id);
-        try {
-            $gloader = new GoogleDataLoader($optionsManager);
-            $gloader->setMode(GoogleAPI::MODE_SERVICE);
-            $gloader->setUser($user->id);
-            $sites = $gloader->listGSCSites(function ($site) {
-                $lastElement = GscTask::selectRaw("MAX(date) as date")
-                    ->where("user_id", $site->user_id)
-                    ->where("site_id", $site->id)
-                    ->where("status", "!=", "disabled")
-                    ->limit(1)
-                    ->first();
-                $firstElement = GscTask::selectRaw("MIN(date) as date")
-                    ->where("user_id", $site->user_id)
-                    ->where("site_id", $site->id)
-                    ->where("status", "!=", "disabled")
-                    ->limit(1)
-                    ->first();
-                if ($lastElement != null && $lastElement->date != null) {
-                    $site->last_date = $lastElement->date;
-                } else {
-                    $site->last_date = null;
-                }
-                if ($firstElement != null && $firstElement->date != null) {
-                    $site->first_date = $firstElement->date;
-                } else {
-                    $site->first_date = null;
-                }
-            });
-            return view("gscaccounts.index")
+       $sites = $repo->updateAndListSites();
+
+       return view("gscaccounts.index")
                 ->with("sites", $sites);
-        } catch (GoogleNeedConfigFileException $e) {
-            Session::flash("fail", "Необходимо загрузить файл сервис аккаунта Google в разделе 'Настройка доступа'");
-            return view("gscaccounts.index")
-                ->with("sites", array());
-        }
+
     }
 
-    public function load(Request $request, $id)
+    public function delete($id, GscTaskRepository $repo)
     {
         $site = GoogleGscSite::findOrFail($id);
-        $user = Auth::user();
-
-        $logger = new LoadLogger($user->id, "gsc");
-
-        $lastElement = GscTask::selectRaw("MAX(date) as date")
-            ->where("user_id", $user->id)
-            ->where("site_id", $site->id)
-            ->where("status", "!=", "disabled")
-            ->limit(1)
-            ->first();
-        if ($lastElement != null && $lastElement->date != null) {
-            $dates = $this->datesRange($lastElement->date, $site->end_date, 50);
-        } else {
-            $dates = $this->datesRange($site->start_date, $site->end_date, 50);
-        }
-        $count = 0;
-        if ($request->has("last_task_id") && $request->get("last_task_id") > 0) {
-            $lastTaskId = $request->get("last_task_id");
-            $logger->write("В задание $lastTaskId добавлено " . count($dates) . " дат");
-        } else {
-            $lastTaskId = null;
-            $logger->write("Новое задание создано, добавлено " . count($dates) . " дат");
-        }
-
-        if (count($dates) > 0) {
-            foreach ($dates as $date) {
-                $count++;
-                $gscTask = new GscTask();
-                $gscTask->date = $date;
-                $gscTask->user_id = $user->id;
-                $gscTask->site_id = $id;
-                $gscTask->save();
-                if ($lastTaskId == null) {
-                    $lastTaskId = $gscTask->id;
-                }
-                GscLoadJob::dispatch($gscTask)
-                    ->onConnection('database')
-                    ->onQueue("gsc_data");
-            }
-            $site->last_task_id = $lastTaskId;
-            $site->parsent = 0;
-            $site->save();
-        }
-        return response()->json(array(
-            "last_task_id" => $lastTaskId,
-            "count" => $count,
-        ));
-    }
-
-    public function status($id)
-    {
-        $site = GoogleGscSite::findOrFail($id);
-        $result = array(
-            "parsent" => $site->parsent,
-            "site_id" => $id
-        );
-        return response()->json($result);
-    }
-
-    public function delete($id)
-    {
-        $site = GoogleGscSite::findOrFail($id);
-        $user = Auth::user();
         $clickHouse = ClickHousePositions::getInstance();
-        $clickHouse->setUser($user->id);
         $clickHouse->setSite($id);
         $clickHouse->delete();
-        GscTask::where("site_id", $id)
-            ->where("user_id", $user->id)
-            ->update(array("status" => "disabled"));
+
+        $repo->disableAllSiteTasks($site);
+
         return back()->withSuccess("Данные сайта полностью удалены!");
     }
 
-    public function stop($id)
+    public function stop($id, GscTaskRepository $repo)
     {
         $site = GoogleGscSite::findOrFail($id);
-        $user = Auth::user();
-        $last_task_id = $site->last_task_id;
 
-        GscTask::where("user_id", $user->id)
-            ->where("site_id", $id)
-            ->where("id", ">=", $last_task_id)
-            ->where("status", "active")
-            ->update(array("status" => "disabled"));
+        $repo->disableLastTasksGroup($site);
+
         return back()->withSuccess("Загрузка остановлена, загруженные данные удалены!");
-    }
-
-    private function datesRange($first, $last, $count, $step = "+1 day", $output_format = "Y-m-d")
-    {
-        $dates = array();
-        $current = strtotime($first);
-        $last = strtotime($last);
-
-        if ($current == $last) {
-            return array();
-        }
-
-        $counter = 0;
-        while ($current <= $last && $counter++ < $count) {
-            $dates[] = date($output_format, $current);
-            $current = strtotime($step, $current);
-        }
-        array_shift($dates);
-        return $dates;
     }
 }
