@@ -7,8 +7,8 @@ use Illuminate\Support\Facades\Auth;
 use Seriyyy95\WPConnector\Connector as WPConnector;
 
 use App\ClickHouse;
-use App\GoogleAnalyticsSite;
-use App\GoogleGscSite;
+use App\Models\GoogleAnalyticsSite;
+use App\Models\GoogleGscSite;
 
 class RequestController extends Controller
 {
@@ -20,33 +20,12 @@ class RequestController extends Controller
 
     public function index(Request $request)
     {
-        $user_id = Auth::user()->id;
-        $gaSites = GoogleAnalyticsSite::where("user_id", $user_id)->get();
-        $gscSites = GoogleGscSite::where("user_id", $user_id)->get();
+        $gaSites = GoogleAnalyticsSite::all();
+        $gscSites = GoogleGscSite::all();
 
         return view("request.index")
             ->with("gaSites", $gaSites)
             ->with("gscSites", $gscSites);
     }
 
-    public function execute(Request $request)
-    {
-        $request->validate([
-            'query' => 'required',
-        ]);
-        $user_id = Auth::user();
-        $query = $request->get("query");
-        try {
-            $rows = ClickHouse::execute($query);
-            return response()->json(array(
-                "data" => $rows,
-                "error" => "",
-            ));
-        } catch (\Exception $e) {
-            return response()->json(array(
-                "data" => array(),
-                "error" => $e->getMessage(),
-            ));
-        }
-    }
 }
